added: added animations and matched it according to weather

+++ weather/lib/models/WeatherModels.dart
@@ -1,13 +1,15 @@
-class Weather
-{
+class Weather {
   final String cityName;
   final double tempreture;
   final String mainCondition;
-  Weather({required this.cityName,required this.mainCondition,required this.tempreture});
-  factory Weather.fromJson(Map<String,dynamic> Json){
+  Weather(
+      {required this.cityName,
+      required this.mainCondition,
+      required this.tempreture});
+  factory Weather.fromJson(Map<String, dynamic> Json) {
     return Weather(
-    cityName: Json['name'],
-    mainCondition: Json['weather'][0]['main'],
-    tempreture: Json['main']['temp'].toDouble());
+        cityName: Json['name'],
+        mainCondition: Json['weather'][0]['main'],
+        tempreture: Json['main']['temp']);
   }
 }
